bot/requirements.txt
fastapi[standard]
uvicorn
python-dotenv
google-genai
pydantic
pydantic-settings
fastapi-users[sqlalchemy]
asyncpg
pgvector
sqlalchemy[ext]